Polars: separate public module
