+++ rust/src/core/mod.rs
@@ -20,11 +20,6 @@
 // Ordering of generic arguments
 // DI, DO, MI, MO, TI, TO, QI, QO
 
-#[cfg(feature = "polars")]
-mod polars;
-#[cfg(feature = "polars")]
-pub use polars::*;
-
 #[cfg(feature = "ffi")]
 mod ffi;
 #[cfg(feature = "ffi")]
