+++ rust/src/lib.rs
@@ -187,3 +187,6 @@ pub mod measures;
 pub mod metrics;
 pub mod traits;
 pub mod transformations;
+
+#[cfg(feature = "polars")]
+pub mod polars;
