+++ rust/src/measurements/make_private_expr/expr_report_noisy_max_gumbel/test.rs
@@ -3,10 +3,10 @@ use polars::prelude::*;
 use polars_arrow::array::{FixedSizeListArray, UInt32Array};
 
 use crate::{
-    core::PrivacyNamespaceHelper,
     error::ErrorVariant,
     measurements::make_private_expr,
     metrics::{PartitionDistance, SymmetricDistance},
+    polars::PrivacyNamespace,
     transformations::expr_discrete_quantile_score::test::get_quantile_test_data,
 };
