+++ rust/src/transformations/make_stable_expr/expr_discrete_quantile_score/plugin_dq_score.rs
@@ -23,7 +23,7 @@ use pyo3_polars::derive::polars_expr;
 use serde::{Deserialize, Serialize};
 
 use crate::{
-    core::OpenDPPlugin, measurements::expr_index_candidates::Candidates, traits::RoundCast,
+    measurements::expr_index_candidates::Candidates, polars::OpenDPPlugin, traits::RoundCast,
 };
 
 use super::{series_to_vec, DQ_SCORE_PLUGIN_NAME};
