+++ rust/src/transformations/make_stable_expr/expr_lit/mod.rs
@@ -2,9 +2,10 @@ use polars_plan::dsl::Expr;
 use polars_plan::logical_plan::LiteralValue;
 use polars_plan::utils::expr_output_name;
 
-use crate::core::{ExprFunction, Function, MetricSpace, StabilityMap, Transformation};
+use crate::core::{Function, MetricSpace, StabilityMap, Transformation};
 use crate::domains::{AtomDomain, ExprDomain, LogicalPlanDomain, Null, OuterMetric, SeriesDomain};
 use crate::error::Fallible;
+use crate::polars::ExprFunction;
 use crate::transformations::DatasetMetric;
 
 #[cfg(test)]
